+++ frontend/src/App.js
@@ -1,7 +1,8 @@
 import React, { useState, useEffect } from "react";
 import Navbar from "./Navbar";
-import Week from "./Week";
-import DetailsPage from "./DetailsPage";
+// import Week from "./Week";
+// import DetailsPage from "./DetailsPage";
+import SignUp from "./SignUp";
 import "./App.css";
 
 const App = () => {
@@ -17,13 +18,14 @@ const App = () => {
 
   return (
     <div className="bg-light">
-      {console.log(weeks)}
+      {/* {console.log(weeks)} */}
       <Navbar />
+      <SignUp />
       {weeks.map((week, index) => {
         return (
           <div key={index}>
-            <Week week={week} />
-            <DetailsPage week={week} />
+            {/* <Week week={week} /> */}
+            {/* <DetailsPage week={week} /> */}
           </div>
         );
       })}
