--- frontend/src/Form.css
@@ -1,32 +1,0 @@
-.form {
-  display: flex;
-  flex-direction: column;
-}
-.input {
-  height: 25px;
-  width: 500px;
-}
-.dropdown {
-  height: 30px;
-  width: 508px;
-}
-.comment {
-  height: 100px;
-  width: 500px;
-}
-.button {
-  float: right;
-  background-color: yellow;
-  border: 1px solid yellow;
-  height: 30px;
-}
-.help {
-  color: red;
-}
-
-input:focus::placeholder {
-  color: transparent;
-}
-textarea:focus::placeholder {
-  color: transparent;
-}
