+++ frontend/src/LoginFormValidationRules.js
@@ -5,7 +5,7 @@ export default function validate(values) {
   } else if (values.name.length < 6) {
     errors.name = "Full name has to be more than 6 characters";
   } else if (!/[A-Z][a-zA-Z]/.test(values.name)) {
-    errors.name = "Name is invalid ";
+    errors.name = "Name is invalid";
   }
   if (!values.email) {
     errors.email = "Email is Required";
@@ -13,7 +13,7 @@ export default function validate(values) {
     errors.email = "Email is invalid";
   }
   if (!values.volunteer) {
-    errors.volunteer = "Select an Option";
+    errors.volunteer = "Please select an Option";
   }
 
   return errors;
