--- frontend/src/SignUpButton.js
@@ -1,13 +1,0 @@
-import React from "react";
-
-const SignUpButton = () => {
-  return (
-    <div>
-      <button type="submit" className="button is-block is-info is-fullwidth">
-        Submit
-      </button>
-    </div>
-  );
-};
-
-export default SignUpButton;
