--- frontend/src/Title.js
@@ -1,5 +1,0 @@
-import React from 'react';
-   const Title =()=>{
-       return <h5>Sign Up form</h5>
-   }
-export default Title
